change table history
